Billing Latest

+++ App_Layer/BillingDetails.cs
@@ -14,11 +14,11 @@ namespace App_Layer
         }
         public int Id { get; set; }
         public string ServiceName { get; set; }
-        public float Qty { get; set; }
-        public float UnitPrize { get; set; }
-        public float Bill { get; set; }
-        public float Paid { get; set; }
-        public float Balance { get; set; }
+        public double Qty { get; set; }
+        public double UnitPrize { get; set; }
+        public double Bill { get; set; }
+        public double Paid { get; set; }
+        public double Balance { get; set; }
         public string CasePaperNo { get; set; }
         public string HospitalId { get; set; }
         public int PatientId { get; set; }
@@ -28,10 +28,10 @@ namespace App_Layer
         public int ModifideBy { get; set; }
         public bool IsActive { get; set; }
         public int QueueId { get; set; }
-        public float TotalBill { get; set; }
-        public float TotalPaid { get; set; }
-        public float TotalBalance { get; set; }
-        public float Total { get; set; }
+        public double TotalBill { get; set; }
+        public double TotalPaid { get; set; }
+        public double TotalBalance { get; set; }
+        public double Total { get; set; }
 
         public List<BillingDetails> lst { get; set; }
     }
